# Bit hacks to improve performance for 1.13 environmental lighting