### Identify the Bug or Feature request

Addresses #4031

### Description of the Change

The `LightingComposite.Blender` implementations have been rewritten to avoid some expensive operations, and to make them more appealing to auto-vectorization on JDK 17.0.1. This also involved removing `MAX_DARKNESS_BOOST_PER_128` as it ended up being 128 anyways and was just extra cruft.

Once the JVM optimizes, the same scenario as in the #4031 has these timings:
```
Timer ZoneRenderer.renderZone (66 elements)
    0.      60 ms  paintComponent
   36.      47 ms  lights
   38.      46 ms  renderLights:renderLightOverlay
   41.      35 ms  renderLightOverlay:drawLights
   42.      35 ms  renderLightOverlay:fillLight
   43.      10 ms  renderLightOverlay:drawBuffer
```
So mainly the improvement is coming from an improved `ScreenBlender`, while the `ConstrainedBrightenBlender` (represented in `renderLightOverlay:drawBuffer`) is just a little more consistent rather than seeing an large average improvement.

### Possible Drawbacks

The code is less obvious now, which does make maintenace / browsing more difficult.

Performance is hardware dependent. While these changes shouldn't given a performance regression anywhere, it's hard to say for sure.

### Documentation Notes

N/A

### Release Notes

- Improve environmental lighting performance.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/RPTools/maptool/4042)
<!-- Reviewable:end -->